Move addToFav and removeFromFav method to player store
