+++ src/stores/player.js
@@ -1,4 +1,13 @@
 import { defineStore } from "pinia";
+import { auth, db } from "@/includes/firebase";
+import {
+  arrayRemove,
+  arrayUnion,
+  doc,
+  getDoc,
+  setDoc,
+  updateDoc,
+} from "firebase/firestore";
 import { Howl } from "howler";
 import helper from "@/includes/helper";
 
@@ -8,6 +17,7 @@ export default defineStore("player", {
       albumId: "",
       artist: "Artist",
       id: "",
+      inFavorites: false,
       picture: "",
       title: "Song title",
     },
@@ -24,7 +34,33 @@ export default defineStore("player", {
     volume: 100,
   }),
   actions: {
+    async addToFav(song) {
+      const favoritesRef = doc(db, "favorites", auth.currentUser.uid);
+      const favoritesSnapshot = await getDoc(favoritesRef);
+
+      if (favoritesSnapshot.exists()) {
+        await updateDoc(favoritesRef, {
+          songs: arrayUnion({ id: song.id }),
+        });
+      } else {
+        await setDoc(favoritesRef, {
+          songs: [{ id: song.id }],
+        });
+      }
+
+      song.inFavorites = true;
+    },
+
+    async removeFromFav(song) {
+      await updateDoc(doc(db, "favorites", auth.currentUser.uid), {
+        songs: arrayRemove({ id: song.id }),
+      });
+
+      song.inFavorites = false;
+    },
+
     async newSong(song) {
+      console.log(song);
       // If song is playing, return
       if (this.currentSong.id === song.id && this.playing) return;
       // If current song is paused
@@ -40,6 +76,7 @@ export default defineStore("player", {
         albumId: song.album_id,
         artist: song.artist,
         id: song.id,
+        inFavorites: song.inFavorites,
         picture: song.picture,
         title: song.title,
       };
